requirements.txt
pygithub
nltk
google-api-python-client
oauth2client
sqlite3
matplotlib
jinja2
retrying